refactor: emit event from filter side-bar
By way of the parent and child relationship between `OffersComponent`
and `SideBarComponent`, emit an event in response to any change in
selected memory filters.

Instead of emitting whole form group members, as in the Signal
implementation from before, pass in only the keys (e.g., 8, 16, 32)
of the selected memory capacities.